refactor: 테스트를 위한 fake 객체의 폴더 위치 변경
