+++ src/test/java/domain/janggi/service/JanggiServiceTest.java
@@ -7,8 +7,8 @@ import domain.board.BoardPosition;
 import domain.piece.Jju;
 import domain.piece.Team;
 import domain.turn.Turn;
-import infra.repository.InMemoryBoardRepository;
-import infra.repository.InMemoryTurnRepository;
+import fake.InMemoryBoardRepository;
+import fake.InMemoryTurnRepository;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
