+++ src/test/java/infra/repository/InMemoryTurnRepositoryTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
 import domain.piece.Team;
 import domain.turn.Turn;
+import fake.InMemoryTurnRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
